+++ Kernel/System/Encode.pm
@@ -15,6 +15,7 @@ use Encode;
 use Encode::Locale;
 use IO::Interactive qw(is_interactive);
 
+our $ObjectManagerIsSingleton = 1;
 our @ObjectDependencies = ();
 
 =head1 NAME