+++ scripts/test/ObjectManager/Singleton.pm
@@ -6,24 +6,20 @@
 # did not receive this file, see http://www.gnu.org/licenses/agpl.txt.
 # --
 
-package scripts::test::ObjectManager::Disabled;    ## no critic
+package scripts::test::ObjectManager::Singleton;    ## no critic
 
 use strict;
 use warnings;
 
-our $ObjectManagerDisabled = 1;
+use Kernel::System::ObjectManager;
+
+our $ObjectManagerIsSingleton = 1;
 our @ObjectDependencies = ();
 
 sub new {
     my ( $Class, %Param ) = @_;
 
     bless \%Param, $Class;
-}
-
-sub Data {
-    my ($Self) = @_;
-
-    return $Self->{Data};
 }
 
 1;